Add makefile and some scripts
- rebrand debug in fake : https://github.com/jsdidierlaurent/monitowall/issues/20
- add Makefile : https://github.com/jsdidierlaurent/monitowall/issues/1
