+++ .gitignore
@@ -7,13 +7,13 @@ vendor
 *.dll
 *.so
 *.dylib
-output
+/build/monitowall*
 
 # Test binary, build with `go test -c`
 *.test
 
 # Output of the go coverage tool, specifically when used with LiteIDE
-*.out
+coverage.txt
 
 # Intellij
 .idea
