+++ cmd/monitowall/monitowall.go
@@ -1,24 +1,17 @@
 package main
 
 import (
-	"github.com/jsdidierlaurent/monitowall/configs"
+	"github.com/jsdidierlaurent/monitowall/config"
 	"github.com/jsdidierlaurent/monitowall/service"
 )
 
-var (
-	GitCommit, Version, BuildTime, OS, Arch string
-)
-
 func main() {
-	// Setup BuildInfo struct
-	buildInfo := configs.InitBuildInfo(GitCommit, Version, BuildTime, OS, Arch)
-
 	// Load Config from File/Env
-	config, err := configs.InitConfig()
+	config, err := config.InitConfig()
 	if err != nil {
 		panic(err)
 	}
 
 	// Start Service
-	service.Start(config, buildInfo)
+	service.Start(config)
 }
