+++ models/info.go
@@ -1,13 +1,30 @@
 package models
 
 import (
-	"github.com/jsdidierlaurent/monitowall/configs"
+	"github.com/jsdidierlaurent/monitowall/config"
 )
 
 type (
 	//InfoResponse response for info route
 	InfoResponse struct {
-		BuildInfo configs.BuildInfo `json:"build-info"`
-		Config    configs.Config    `json:"configuration"`
+		BuildInfo BuildInfo     `json:"build-info"`
+		Config    config.Config `json:"configuration"`
+	}
+
+	BuildInfo struct {
+		Version   string `json:"version"`
+		GitCommit string `json:"git-commit"`
+		BuildTime string `json:"build-time"`
 	}
 )
+
+func NewInfoResponse(version string, gitCommit string, buildTime string, config *config.Config) *InfoResponse {
+	return &InfoResponse{
+		BuildInfo: BuildInfo{
+			Version:   version,
+			GitCommit: gitCommit,
+			BuildTime: buildTime,
+		},
+		Config: *config,
+	}
+}
