+++ monitorable/ping/model/ping_params.go
@@ -1,3 +1,5 @@
+//+build !faker
+
 package model
 
 type (
