+++ monitorable/ping/model/ping_params_faker.go
@@ -1,0 +1,17 @@
+//+build faker
+
+package model
+
+import "github.com/jsdidierlaurent/monitowall/models/tiles"
+
+type (
+	PingParams struct {
+		Hostname string           `json:"hostname" query:"hostname"`
+		Status   tiles.TileStatus `json:"status" query:"status"`
+		Message  string           `json:"message" query:"message"`
+	}
+)
+
+func (p *PingParams) Validate() bool {
+	return p.Hostname != ""
+}
