+++ monitorable/ping/usecase.go
@@ -1,13 +1,17 @@
 package ping
 
 import (
-	"github.com/jsdidierlaurent/monitowall/models/tiles"
+	. "github.com/jsdidierlaurent/monitowall/models/tiles"
 	"github.com/jsdidierlaurent/monitowall/pkg/bind"
+)
+
+const (
+	PingTileSubType TileSubType = "PING"
 )
 
 // Usecase represent the ping's usecases
 type (
 	Usecase interface {
-		Ping(binder bind.Binder) (*tiles.HealthTile, error)
+		Ping(binder bind.Binder) (*HealthTile, error)
 	}
 )
