+++ monitorable/ping/usecase/ping_usecase_test.go
@@ -6,6 +6,7 @@ import (
 	"time"
 
 	"github.com/jsdidierlaurent/monitowall/models/tiles"
+	. "github.com/jsdidierlaurent/monitowall/monitorable/ping"
 	"github.com/jsdidierlaurent/monitowall/monitorable/ping/mocks"
 	"github.com/jsdidierlaurent/monitowall/monitorable/ping/model"
 	pkgMock "github.com/jsdidierlaurent/monitowall/pkg/bind/mocks"
