+++ scripts/install
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+
+echo " - Installing tools ..."
+go get gotest.tools/gotestsum
+go mod tidy
+
+echo " - Installing module dependencies ..."
+go mod download
